Move remaining kinesis-decoder
