+++ lib/decoder/bytes-lambda-event.test.json
