+++ lib/decoder/bytes-lambda.test.js
@@ -3,8 +3,6 @@ const bytesLambda = require('./bytes-lambda')
 const { BadEventError } = require('../errors')
 
 describe('bytes-lambda', () => {
-  afterEach(() => jest.restoreAllMocks())
-
   describe('#detect', () => {
     it('detects gzipped data', async () => {
       expect(bytesLambda.detect(zlib.gzipSync(''))).toEqual(true)
